cleanup BFAdapter streaming tests
